backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
asyncpg>=0.29.0
redis>=5.0.0
httpx>=0.26.0
pydantic[email]>=2.5.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0
stripe>=7.0.0
sqlalchemy[asyncio]>=2.0.0
neo4j>=5.0.0
ollama>=0.1.0
structlog>=24.0.0
aiohttp
anthropic>=0.40.0
bcrypt>=4.0.0
sentry-sdk[fastapi]>=1.0.0
pgvector>=0.2.0
